web/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
itsdangerous>=2.2.0
psycopg2-binary>=2.9.9
fpdf2>=2.7.9
